Rendu4/backend/requirements.txt
psycopg2==2.9.9
python-dotenv==1.0.0
Flask==3.0.0
Flask-JWT-Extended==4.6.0
flasgger==0.9.7.1
flask-cors==4.0.0